rename: train.py -> simulation_objects.py
